backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
PyMuPDF>=1.23.0
python-docx>=1.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
google-genai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
pandas>=2.0.0